+++ HISTORY.rst
@@ -168,6 +168,8 @@
   - anonymous: 5 requests per second
   - logged users: 10 requests per second
 
+- Add support for deleting workflow instances via the UI and the API.
+
 4.9.6 (XXXX-XX-XX)
 ==================
 - Rename file metadata model mixin class for clarity.